Refactored code. Moved the bot commands into separate files for better readability.
