+++ commands/all.ts
@@ -1,0 +1,4 @@
+import "./start"
+import "./chatwith"
+import "./greeting"
+import "./language"
