+++ scripts/start.ts
@@ -1,4 +1,5 @@
 import { bot } from '../src/bot'
+import "../commands/all"
 
 bot.catch(console.error)
