+++ src/bot.ts
@@ -1,11 +1,8 @@
-import { Bot, Context, InlineKeyboard, session, SessionFlavor } from 'grammy'
+import { Bot, session } from 'grammy'
 import dotenv from "dotenv"
-import { I18n, I18nFlavor } from "@grammyjs/i18n";
+import { I18n } from "@grammyjs/i18n";
+import { MyContext } from '../types/myContext';
 
-interface SessionData {
-    __language_code?: string;
-}
-type MyContext = Context & SessionFlavor<SessionData> & I18nFlavor;
 
 dotenv.config()
 const {
@@ -30,55 +27,3 @@ bot.use(
 
 bot.use(i18n);
 
-async function setLocalizedCommands(ctx: MyContext) {
-    await ctx.api.setMyCommands([
-        {
-            command: "start",
-            description: ctx.t("commandStart"),
-        },
-        {
-            command: "greeting",
-            description: ctx.t("commandGreeting"),
-        },
-        {
-            command: "language",
-            description: ctx.t("commandLanguage"),
-        },
-    ]);
-}
-
-
-bot.command("start", async (ctx) => {
-    await setLocalizedCommands(ctx)
-    await ctx.reply(ctx.t("startText"))
-})
-bot.command("greeting", async (ctx) => {
-    const admin = 1915118825
-    if (ctx.from?.id == admin) {
-        await ctx.reply(ctx.t("greetingRonText"))
-    }
-    else {
-        await ctx.reply(`${ctx.t("greetingText")} ${ctx.from?.first_name}`)
-    }
-})
-
-
-
-const langKeyboard = new InlineKeyboard().text("Русский ", "ru").text("English", "en")
-
-bot.command("language", async (ctx) => {
-    await ctx.reply(ctx.t("langText"), {
-        reply_markup: langKeyboard
-    })
-})
-
-bot.callbackQuery(["en", "ru"], async ctx => {
-    await ctx.answerCallbackQuery(ctx.t("lng"))
-    await ctx.i18n.setLocale(ctx.callbackQuery.data);
-    await ctx.deleteMessage()
-    await ctx.reply(ctx.t("langSetComm"))
-    await setLocalizedCommands(ctx);
-})
-bot.on('message', async ctx => {
-    await ctx.reply(ctx.message.text!)
-})
